requirements.txt
# Crawlers
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.3
playwright>=1.49          # for JS-heavy shop pages

# Storage
duckdb>=1.1
psycopg2-binary>=2.9

# DBT
dbt-core>=1.9
dbt-duckdb>=1.9

# Utilities
tqdm>=4.66
python-dotenv>=1.0
pydantic>=2.10
schedule>=1.2             # periodic crawl scheduling
